Small changes to MockHttp
