+++ tests/Tests.MockHttp/Tests.MockHttp.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFrameworks>netcoreapp3.1;net48</TargetFrameworks>
+    <TargetFrameworks>netcoreapp3.1;net48;net5.0</TargetFrameworks>
     <IsPackable>false</IsPackable>
   </PropertyGroup>
